add:model annotation index schema
